Move iframe.resize parser next to its payload type and fold frame styles into CollectionGrid
The review thumbnail message layer had a renaming re-export module
(iframeResizeMessage.ts) that gave the shared iframe.resize wire protocol a
second vocabulary. The parser now lives in shared/constants/iframe-resize.ts
next to the payload type it validates, and the frame-style helpers live in
CollectionGrid next to the Frame styled component that consumes their CSS
variables. Also drops the tautological THUMBNAIL_BOOTSTRAP_SCALE constant test.
