+++ code/core/src/manager/components/review/components/CollectionGrid.tsx
@@ -1,14 +1,13 @@
-import React, { type FC } from 'react';
+import React, { type CSSProperties, type FC } from 'react';
 
 import { Badge, Button, Loader } from 'storybook/internal/components';
 import { styled } from 'storybook/theming';
 
-import { fallbackStoryInfo, type StoryInfo } from '../review-types.ts';
 import {
-  DEFAULT_CONTENT_WIDTH,
-  getPreviewFrameStyle,
-  hasFixedViewportAspect,
-} from './iframeResizeMessage.ts';
+  hasFixedViewportDimensions,
+  type IframeResizeDimensions,
+} from '../../../../shared/constants/iframe-resize.ts';
+import { fallbackStoryInfo, type StoryInfo } from '../review-types.ts';
 import { usePreviewThumbnail } from './usePreviewThumbnail.ts';
 
 // Per-breakpoint grid: `cols` columns (each cell clamped to 400px) capped at
@@ -59,6 +58,37 @@ const FrameShell = styled.div({
   aspectRatio: '3 / 2',
   position: 'relative',
 });
+
+/** Default `--content-w` before the embed iframe reports its size. */
+const DEFAULT_CONTENT_WIDTH = 300;
+
+/** Pre-measurement scale so the embed iframe viewport is 2× the frame width (100% / 0.5). */
+const THUMBNAIL_BOOTSTRAP_SCALE = 0.5;
+
+// Feeds the CSS variables consumed by `Frame` below (`--scale`, `--content-w`,
+// `--vp-w`/`--vp-h`); `viewportFill` toggles the `data-viewport-fill` branch.
+const getPreviewFrameLayout = (
+  dimensions: IframeResizeDimensions | null
+): { style: CSSProperties; viewportFill: boolean } => {
+  if (!dimensions) {
+    return {
+      style: { '--scale': THUMBNAIL_BOOTSTRAP_SCALE } as CSSProperties,
+      viewportFill: false,
+    };
+  }
+
+  if (hasFixedViewportDimensions(dimensions.viewport)) {
+    return {
+      style: {
+        '--vp-w': dimensions.viewport.width,
+        '--vp-h': dimensions.viewport.height,
+      } as CSSProperties,
+      viewportFill: true,
+    };
+  }
+
+  return { style: { '--content-w': dimensions.width } as CSSProperties, viewportFill: false };
+};
 
 const Frame = styled.a(({ theme }) => ({
   position: 'absolute',
@@ -215,9 +245,7 @@ const StoryPreviewCell: FC<{
   } = usePreviewThumbnail({ storyId, getPreviewHref, previewsPaused });
 
   const { component, name } = deriveStoryInfo(info);
-  const viewport = rememberedDimensions?.viewport;
-  const viewportFill = hasFixedViewportAspect(viewport);
-  const frameStyle = getPreviewFrameStyle(rememberedDimensions);
+  const { style: frameStyle, viewportFill } = getPreviewFrameLayout(rememberedDimensions);
 
   const preview = src ? (
     <div data-preview-scale>
